requirements.txt
numpy
pyqt>=5.0
nilearn
nipype
jinja2
matplotlib
